Extract NetConfig created from Json to a separate class
