+++ ethereumj-core/src/main/java/org/ethereum/config/net/AbstractNetConfig.java
@@ -39,64 +39,4 @@ public class AbstractNetConfig implements BlockchainNetConfig {
         // TODO make a guard wrapper which throws exception if the requested constant differs among configs
         return configs[0].getConstants();
     }
-
-    /**
-     * Convert JSON config from genesis to Java blockchain net config.
-     */
-    public static BlockchainNetConfig fromGenesisConfig(Map<String, String> config) throws RuntimeException {
-        AbstractNetConfig netConfig = new AbstractNetConfig();
-
-        final List<String> keys = Arrays.asList("homesteadBlock", "daoForkBlock", "EIP150Block", "EIP155Block", "EIP158Block");
-        final ArrayList<Integer> orderedBlocks = new ArrayList<>();
-        final HashMap<Integer, String> blockToKey = new HashMap<>();
-
-        // #1 create block number to config map
-        for (String key : keys) {
-            Integer blockNumber = getInteger(config, key);
-            if (blockNumber != null) {
-                if (orderedBlocks.contains(blockNumber)) {
-                    throw new RuntimeException("Genesis net config contains duplicate blocks");
-                }
-                orderedBlocks.add(blockNumber);
-                blockToKey.put(blockNumber, key);
-            }
-        }
-
-        // #2 sort block numbers before adding configs
-        Collections.sort(orderedBlocks);
-
-        // hardcoded initial config
-        Integer prevBlockNumber = 0;    // frontier block
-        netConfig.add(0, new FrontierConfig());
-
-        // #3 fill configs in proper order
-        for (Integer blockNumber : orderedBlocks) {
-            String key = blockToKey.get(blockNumber);
-            switch (key) {
-                case "homesteadBlock":
-                    netConfig.add(blockNumber, new HomesteadConfig());
-                    break;
-                case "daoForkBlock":
-                    if ("true".equalsIgnoreCase(config.get("daoForkSupport"))) {
-                        netConfig.add(blockNumber, new DaoHFConfig().withForkBlock(blockNumber));
-                    } else {
-                        netConfig.add(blockNumber, new DaoNoHFConfig().withForkBlock(blockNumber));
-                    }
-                    break;
-                case "EIP150Block":
-                    netConfig.add(blockNumber, new Eip150HFConfig(netConfig.getConfigForBlock(prevBlockNumber)));
-                    break;
-                // TODO handle other EIP configs
-                default:
-                    continue;
-            }
-            prevBlockNumber = blockNumber;
-        }
-
-        return netConfig;
-    }
-
-    private static Integer getInteger(Map<String, String> config, String key) {
-        return config.containsKey(key) ? Integer.parseInt(config.get(key)) : null;
-    }
 }
